Properly ignore public/* as recommended by Gatsby
As a result of this, I've moved all assets from public/static
to src/assets, and they are now loaded via webpack.

There's some weirdness with loading fender-styles.css, it's currently being
imported in html.js, pages/index.js, as well as via `public/styles.css` when
building for production. It's a bit weird, but it works, and I figure we're
switching to glamorous at some point so I'm not too worried for now.